Agrupacion roles al registrar y correcion BD
