+++ src/main/java/co/edu/unicauca/academicprojectsprototype/main/Main.java
@@ -9,6 +9,8 @@ import co.edu.unicauca.academicprojectsprototype.access.ICompanyRepository;
 import co.edu.unicauca.academicprojectsprototype.access.IStudentRepository;
 import co.edu.unicauca.academicprojectsprototype.domain.services.CompanyService;
 import co.edu.unicauca.academicprojectsprototype.domain.services.ICompanyService;
+import co.edu.unicauca.academicprojectsprototype.domain.services.IStudentService;
+import co.edu.unicauca.academicprojectsprototype.domain.services.StudentService;
 import co.edu.unicauca.academicprojectsprototype.presentation.GUIHomeCoordinator;
 import co.edu.unicauca.academicprojectsprototype.presentation.GUIHomeWithoutLog;
 import javax.swing.JFrame;
@@ -22,10 +24,13 @@ public class Main {
         
         // Hacia futuro el tipo de repositorio lo podemos leer de un archivo plano
         // setup.properties, asi no tendriamos que recompilar la aplicación
-        ICompanyRepository repository = Factory.getInstance().getRepository("SQLITE");// Podria ir ARRAYS/SQLITE
-        ICompanyService service = new CompanyService(repository);
-
-        GUIHomeWithoutLog instance = new GUIHomeWithoutLog(service);
+        ICompanyRepository repositoryCompany = Factory.getInstance().getRepositoryCompany("SQLITE");// Podria ir ARRAYS/SQLITE
+        IStudentRepository repositoryStudent = Factory.getInstance().getRepositoryStudent("SQLITE");
+        
+        ICompanyService serviceCompany = new CompanyService(repositoryCompany);
+        IStudentService serviceStudent = new StudentService(repositoryStudent);
+        
+        GUIHomeWithoutLog instance = new GUIHomeWithoutLog(serviceCompany,serviceStudent);
         instance.setExtendedState(JFrame.MAXIMIZED_BOTH);
         instance.setVisible(true);
     }
