+++ src/main/java/co/edu/unicauca/academicprojectsprototype/presentation/GUINewProject.java
@@ -25,6 +25,7 @@ public class GUINewProject extends javax.swing.JFrame {
     @SuppressWarnings("unchecked")
     // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
     private void initComponents() {
+        java.awt.GridBagConstraints gridBagConstraints;
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
