--- src/main/java/co/edu/unicauca/academicprojectsprototype/presentation/GUINewStudent.java
@@ -1,359 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
- */
-package co.edu.unicauca.academicprojectsprototype.presentation;
-
-import co.edu.unicauca.academicprojectsprototype.domain.entities.Company;
-import co.edu.unicauca.academicprojectsprototype.domain.entities.Student;
-import co.edu.unicauca.academicprojectsprototype.domain.services.ICompanyService;
-import co.edu.unicauca.academicprojectsprototype.domain.services.IStudentService;
-import co.edu.unicauca.academicprojectsprototype.domain.services.validateStudent;
-import co.edu.unicauca.academicprojectsprototype.infra.Messages;
-
-/**
- *
- * @author anvig
- */
-public class GUINewStudent extends javax.swing.JFrame {
-    private IStudentService studentService;
-    /**
-     * Creates new form NewStudent
-     */
-    public GUINewStudent(IStudentService service) {
-        this.studentService = service;
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-        java.awt.GridBagConstraints gridBagConstraints;
-
-        jPHead = new javax.swing.JPanel();
-        lbTitleProyect = new java.awt.Label();
-        jPTittleNewUser = new javax.swing.JPanel();
-        jLTittleNewStudent = new javax.swing.JLabel();
-        jPContent = new javax.swing.JPanel();
-        jLNameStudent = new javax.swing.JLabel();
-        jFieldNameStudent = new javax.swing.JTextField();
-        jLCodeStudent = new javax.swing.JLabel();
-        jFieldCodeStudent = new javax.swing.JTextField();
-        jLTelStudent = new javax.swing.JLabel();
-        jFieldTelStudent = new javax.swing.JTextField();
-        jLEmailStudent = new javax.swing.JLabel();
-        jFieldEmailStudent = new javax.swing.JTextField();
-        jLPasswordStudent = new javax.swing.JLabel();
-        jFieldPassWordStudent = new javax.swing.JPasswordField();
-        jBtnSaveStudent = new javax.swing.JButton();
-        jCheckSeePass = new javax.swing.JCheckBox();
-        jPButtom = new javax.swing.JPanel();
-        jBtnBackSelectUser = new javax.swing.JButton();
-
-        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
-
-        jPHead.setBackground(new java.awt.Color(236, 230, 240));
-        jPHead.setForeground(new java.awt.Color(0, 0, 0));
-        jPHead.setLayout(new javax.swing.BoxLayout(jPHead, javax.swing.BoxLayout.LINE_AXIS));
-
-        lbTitleProyect.setAlignment(java.awt.Label.CENTER);
-        lbTitleProyect.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
-        lbTitleProyect.setText("Sistema gestor de proyectos academicos");
-        jPHead.add(lbTitleProyect);
-
-        jPTittleNewUser.setBackground(new java.awt.Color(255, 255, 255));
-
-        jLTittleNewStudent.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
-        jLTittleNewStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLTittleNewStudent.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
-        jLTittleNewStudent.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/NewUser.png"))); // NOI18N
-        jLTittleNewStudent.setText("Nuevo estudiante");
-        jLTittleNewStudent.setHorizontalTextPosition(javax.swing.SwingConstants.RIGHT);
-        jPTittleNewUser.add(jLTittleNewStudent);
-
-        jPContent.setBackground(new java.awt.Color(236, 230, 240));
-        jPContent.setForeground(new java.awt.Color(0, 0, 0));
-        jPContent.setLayout(new java.awt.GridBagLayout());
-
-        jLNameStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jLNameStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLNameStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        jLNameStudent.setText("*Nombre:");
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 0;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jLNameStudent, gridBagConstraints);
-
-        jFieldNameStudent.setBackground(new java.awt.Color(255, 255, 255));
-        jFieldNameStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jFieldNameStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jFieldNameStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
-        jFieldNameStudent.setText("Julian Lopez");
-        jFieldNameStudent.setMaximumSize(new java.awt.Dimension(500, 100));
-        jFieldNameStudent.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jFieldNameStudentActionPerformed(evt);
-            }
-        });
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 1;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jFieldNameStudent, gridBagConstraints);
-
-        jLCodeStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jLCodeStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLCodeStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        jLCodeStudent.setText("Codigo:");
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 2;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jLCodeStudent, gridBagConstraints);
-
-        jFieldCodeStudent.setBackground(new java.awt.Color(255, 255, 255));
-        jFieldCodeStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jFieldCodeStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jFieldCodeStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
-        jFieldCodeStudent.setText("1046123");
-        jFieldCodeStudent.setMaximumSize(new java.awt.Dimension(500, 100));
-        jFieldCodeStudent.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jFieldCodeStudentActionPerformed(evt);
-            }
-        });
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 3;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jFieldCodeStudent, gridBagConstraints);
-
-        jLTelStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jLTelStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLTelStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        jLTelStudent.setText("Telefono de contacto:");
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 4;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jLTelStudent, gridBagConstraints);
-
-        jFieldTelStudent.setBackground(new java.awt.Color(255, 255, 255));
-        jFieldTelStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jFieldTelStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jFieldTelStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
-        jFieldTelStudent.setText("317123");
-        jFieldTelStudent.setMaximumSize(new java.awt.Dimension(500, 100));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 5;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jFieldTelStudent, gridBagConstraints);
-
-        jLEmailStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jLEmailStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLEmailStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        jLEmailStudent.setText("Email:");
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 6;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jLEmailStudent, gridBagConstraints);
-
-        jFieldEmailStudent.setBackground(new java.awt.Color(255, 255, 255));
-        jFieldEmailStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jFieldEmailStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jFieldEmailStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
-        jFieldEmailStudent.setText("lopez@gmail.com");
-        jFieldEmailStudent.setMaximumSize(new java.awt.Dimension(500, 100));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 7;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jFieldEmailStudent, gridBagConstraints);
-
-        jLPasswordStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jLPasswordStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jLPasswordStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
-        jLPasswordStudent.setText("Clave:");
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 8;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jLPasswordStudent, gridBagConstraints);
-
-        jFieldPassWordStudent.setBackground(new java.awt.Color(255, 255, 255));
-        jFieldPassWordStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
-        jFieldPassWordStudent.setForeground(new java.awt.Color(0, 0, 0));
-        jFieldPassWordStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
-        jFieldPassWordStudent.setText("jPasswordField1");
-        jFieldPassWordStudent.setMaximumSize(new java.awt.Dimension(500, 100));
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 9;
-        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
-        gridBagConstraints.insets = new java.awt.Insets(5, 0, 0, 0);
-        jPContent.add(jFieldPassWordStudent, gridBagConstraints);
-
-        jBtnSaveStudent.setBackground(new java.awt.Color(0, 0, 0));
-        jBtnSaveStudent.setFont(new java.awt.Font("SansSerif", 0, 14)); // NOI18N
-        jBtnSaveStudent.setForeground(new java.awt.Color(255, 255, 255));
-        jBtnSaveStudent.setText("GUARDAR");
-        jBtnSaveStudent.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtnSaveStudentActionPerformed(evt);
-            }
-        });
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridy = 10;
-        gridBagConstraints.insets = new java.awt.Insets(10, 0, 0, 0);
-        jPContent.add(jBtnSaveStudent, gridBagConstraints);
-
-        jCheckSeePass.setText("Mostrar clave");
-        jCheckSeePass.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jCheckSeePassActionPerformed(evt);
-            }
-        });
-        gridBagConstraints = new java.awt.GridBagConstraints();
-        gridBagConstraints.gridx = 1;
-        gridBagConstraints.gridy = 9;
-        gridBagConstraints.insets = new java.awt.Insets(10, 0, 0, 0);
-        jPContent.add(jCheckSeePass, gridBagConstraints);
-
-        jPButtom.setBackground(new java.awt.Color(236, 230, 240));
-        jPButtom.setLayout(new java.awt.FlowLayout(java.awt.FlowLayout.LEFT));
-
-        jBtnBackSelectUser.setBackground(new java.awt.Color(236, 230, 240));
-        jBtnBackSelectUser.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/back.png"))); // NOI18N
-        jBtnBackSelectUser.setBorderPainted(false);
-        jBtnBackSelectUser.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                jBtnBackSelectUserActionPerformed(evt);
-            }
-        });
-        jPButtom.add(jBtnBackSelectUser);
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
-        getContentPane().setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPTittleNewUser, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(jPHead, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 808, Short.MAX_VALUE)
-            .addComponent(jPButtom, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-            .addComponent(jPContent, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addComponent(jPHead, javax.swing.GroupLayout.PREFERRED_SIZE, 101, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPTittleNewUser, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPContent, javax.swing.GroupLayout.PREFERRED_SIZE, 312, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
-                .addComponent(jPButtom, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-
-        pack();
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void jFieldNameStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldNameStudentActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_jFieldNameStudentActionPerformed
-
-    private void jFieldCodeStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldCodeStudentActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_jFieldCodeStudentActionPerformed
-
-    private void jBtnBackSelectUserActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtnBackSelectUserActionPerformed
-        // TODO add your handling code here:
-        GUISelectRegistUser selecU = new GUISelectRegistUser(null);
-        selecU.setVisible(true);
-        this.dispose();
-    }//GEN-LAST:event_jBtnBackSelectUserActionPerformed
-
-    private void jCheckSeePassActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCheckSeePassActionPerformed
-        // TODO add your handling code here:
-        if (jCheckSeePass.isSelected()) {
-        jFieldPassWordStudent.setEchoChar((char) 0); // Mostrar la contraseña
-    } else {
-        jFieldPassWordStudent.setEchoChar('●'); // Ocultar la contraseña con puntos
-    }
-    }//GEN-LAST:event_jCheckSeePassActionPerformed
-
-    private void jBtnSaveStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtnSaveStudentActionPerformed
-        // TODO add your handling code here:
-        validateStudent validar = new validateStudent();
-        
-        String name = jFieldNameStudent.getText();
-        if (name.equals("")){
-            Messages.showMessageDialog("Debe agregar el nombre", "Atención");
-            jFieldNameStudent.requestFocus();
-            return;
-        }
-      
-        String code = jFieldCodeStudent.getText();
-        if (code.equals("")){
-            Messages.showMessageDialog("Debe agregar el codigo", "Atención");
-            jFieldCodeStudent.requestFocus();
-            return;
-        }
-        
-        String telefono = jFieldTelStudent.getText();
-        
-        String email = jFieldEmailStudent.getText().trim();
-        if (email.equals("")){
-            Messages.showMessageDialog("Debe agregar el Email", "Atención");
-            jFieldEmailStudent.requestFocus();
-            return;
-        }
-        
-        String pass = jFieldPassWordStudent.getText().trim();
-        if (pass.equals("")){
-            Messages.showMessageDialog("Debe agregar la Contraseña", "Atención");
-            jFieldPassWordStudent.requestFocus();
-            return;
-        }
-        if(!validar.validarClave(pass)){
-            validar.mostrarEjemploClave();
-        }
-        if(validar.validarRegistroEstudiante(email, telefono, pass)){
-        studentService.save(new Student(name, code, telefono, email, pass));
-        Messages.showMessageDialog("Estudiante registrado exitosamente", "Atencion");
-        GUIHomeStudent hStudent = new GUIHomeStudent();
-        hStudent.setVisible(true);
-        this.dispose();
-        }
-    }//GEN-LAST:event_jBtnSaveStudentActionPerformed
-
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton jBtnBackSelectUser;
-    private javax.swing.JButton jBtnSaveStudent;
-    private javax.swing.JCheckBox jCheckSeePass;
-    private javax.swing.JTextField jFieldCodeStudent;
-    private javax.swing.JTextField jFieldEmailStudent;
-    private javax.swing.JTextField jFieldNameStudent;
-    private javax.swing.JPasswordField jFieldPassWordStudent;
-    private javax.swing.JTextField jFieldTelStudent;
-    private javax.swing.JLabel jLCodeStudent;
-    private javax.swing.JLabel jLEmailStudent;
-    private javax.swing.JLabel jLNameStudent;
-    private javax.swing.JLabel jLPasswordStudent;
-    private javax.swing.JLabel jLTelStudent;
-    private javax.swing.JLabel jLTittleNewStudent;
-    private javax.swing.JPanel jPButtom;
-    private javax.swing.JPanel jPContent;
-    private javax.swing.JPanel jPHead;
-    private javax.swing.JPanel jPTittleNewUser;
-    private java.awt.Label lbTitleProyect;
-    // End of variables declaration//GEN-END:variables
-}
