+++ src/main/java/co/edu/unicauca/academicprojectsprototype/presentation/GUIRegisteredUser.java
@@ -1,0 +1,1046 @@
+/*
+ * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
+ * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
+ */
+package co.edu.unicauca.academicprojectsprototype.presentation;
+
+import co.edu.unicauca.academicprojectsprototype.domain.entities.Company;
+import co.edu.unicauca.academicprojectsprototype.domain.entities.Sector;
+import co.edu.unicauca.academicprojectsprototype.domain.entities.Student;
+import co.edu.unicauca.academicprojectsprototype.domain.services.ICompanyService;
+import co.edu.unicauca.academicprojectsprototype.domain.services.IStudentService;
+import co.edu.unicauca.academicprojectsprototype.domain.services.validateCompany;
+import co.edu.unicauca.academicprojectsprototype.domain.services.validateStudent;
+import co.edu.unicauca.academicprojectsprototype.infra.Messages;
+import java.awt.CardLayout;
+
+/**
+ *
+ * @author lopez
+ */
+public class GUIRegisteredUser extends javax.swing.JFrame {
+
+    CardLayout cardLayout;
+    private ICompanyService companyService;
+    private IStudentService studentService;
+
+    /**
+     * Creates new form GUIRegisteredUser
+     */
+    public GUIRegisteredUser(ICompanyService serviceCompany, IStudentService serviceStudent) {
+        initComponents();
+        this.companyService = serviceCompany;
+        this.studentService = serviceStudent;
+        fillSectors();
+    }
+
+    private void fillSectors() {
+        jCBSectorCompany.removeAllItems();
+        for (Sector sector : Sector.values()) {
+            jCBSectorCompany.addItem(sector.toString());
+        }
+    }
+
+    /**
+     * This method is called from within the constructor to initialize the form.
+     * WARNING: Do NOT modify this code. The content of this method is always
+     * regenerated by the Form Editor.
+     */
+    @SuppressWarnings("unchecked")
+    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
+    private void initComponents() {
+        java.awt.GridBagConstraints gridBagConstraints;
+
+        RadiobuttonGroupSelecRol = new javax.swing.ButtonGroup();
+        jPHead = new javax.swing.JPanel();
+        lbTitleProyect = new java.awt.Label();
+        jPButtom = new javax.swing.JPanel();
+        jBtnBackSelectUser = new javax.swing.JButton();
+        jPContent = new javax.swing.JPanel();
+        jPSelecRol = new javax.swing.JPanel();
+        jLSelectRol = new javax.swing.JLabel();
+        jRbtnStudent = new javax.swing.JRadioButton();
+        jRbtnCompany = new javax.swing.JRadioButton();
+        jRbtnCoordi = new javax.swing.JRadioButton();
+        jPRolInformation = new javax.swing.JPanel();
+        jPCompanyInformation = new javax.swing.JPanel();
+        jLTittleNewCompany = new javax.swing.JLabel();
+        jScrolInfoCompany = new javax.swing.JScrollPane();
+        jTextAInfoCompany = new javax.swing.JTextArea();
+        jPCoordinatorInformation = new javax.swing.JPanel();
+        jLTittleNewCompany1 = new javax.swing.JLabel();
+        jScrolInfoCoordi = new javax.swing.JScrollPane();
+        jTextAInfoCoordi = new javax.swing.JTextArea();
+        jPStudentInformation = new javax.swing.JPanel();
+        jLTittleNewStudent = new javax.swing.JLabel();
+        jScrolInfoStudent = new javax.swing.JScrollPane();
+        jTextAInfoStudent = new javax.swing.JTextArea();
+        jPRegisterUser = new javax.swing.JPanel();
+        jPCoordiRegister = new javax.swing.JPanel();
+        jPStudentRegister = new javax.swing.JPanel();
+        jLNameStudent = new javax.swing.JLabel();
+        jFieldNameStudent = new javax.swing.JTextField();
+        jLCodeStudent = new javax.swing.JLabel();
+        jFieldCodeStudent = new javax.swing.JTextField();
+        jLTelStudent = new javax.swing.JLabel();
+        jFieldTelStudent = new javax.swing.JTextField();
+        jLEmailStudent = new javax.swing.JLabel();
+        jFieldEmailStudent = new javax.swing.JTextField();
+        jLPasswordStudent = new javax.swing.JLabel();
+        jCheckSeePassStudent = new javax.swing.JCheckBox();
+        jFieldPassWordStudent = new javax.swing.JPasswordField();
+        jBtnSaveStudent = new javax.swing.JButton();
+        jPCompanyRegister = new javax.swing.JPanel();
+        jLNitCompany = new javax.swing.JLabel();
+        jFieldNitCompany = new javax.swing.JTextField();
+        jLNameCompany = new javax.swing.JLabel();
+        jFieldNameCompany = new javax.swing.JTextField();
+        jLTelCompany = new javax.swing.JLabel();
+        jFieldTelCompany = new javax.swing.JTextField();
+        jLWebCompany = new javax.swing.JLabel();
+        jFieldWebCompany = new javax.swing.JTextField();
+        jLSectorCompany = new javax.swing.JLabel();
+        jCBSectorCompany = new javax.swing.JComboBox<>();
+        jLEmailCompany = new javax.swing.JLabel();
+        jFieldEmailCompany = new javax.swing.JTextField();
+        jLPasswordCompany = new javax.swing.JLabel();
+        jFieldPassWordCompany = new javax.swing.JPasswordField();
+        jCheckSeePassCompany = new javax.swing.JCheckBox();
+        jBtnSaveCompany = new javax.swing.JButton();
+        jPTittleNewUser = new javax.swing.JPanel();
+        jLTittleNewUser = new javax.swing.JLabel();
+
+        setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
+
+        jPHead.setBackground(new java.awt.Color(236, 230, 240));
+        jPHead.setForeground(new java.awt.Color(0, 0, 0));
+        jPHead.setLayout(new javax.swing.BoxLayout(jPHead, javax.swing.BoxLayout.LINE_AXIS));
+
+        lbTitleProyect.setAlignment(java.awt.Label.CENTER);
+        lbTitleProyect.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
+        lbTitleProyect.setText("Sistema gestor de proyectos academicos");
+        jPHead.add(lbTitleProyect);
+
+        jPButtom.setBackground(new java.awt.Color(236, 230, 240));
+        jPButtom.setLayout(new java.awt.FlowLayout(java.awt.FlowLayout.LEFT));
+
+        jBtnBackSelectUser.setBackground(new java.awt.Color(236, 230, 240));
+        jBtnBackSelectUser.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/back.png"))); // NOI18N
+        jBtnBackSelectUser.setBorderPainted(false);
+        jPButtom.add(jBtnBackSelectUser);
+
+        jPContent.setBackground(new java.awt.Color(236, 230, 240));
+        jPContent.setForeground(new java.awt.Color(0, 0, 0));
+        jPContent.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        jPSelecRol.setBackground(new java.awt.Color(255, 255, 255));
+        jPSelecRol.setLayout(new org.netbeans.lib.awtextra.AbsoluteLayout());
+
+        jLSelectRol.setText("Seleccione su rol :");
+        jPSelecRol.add(jLSelectRol, new org.netbeans.lib.awtextra.AbsoluteConstraints(110, 10, -1, -1));
+
+        RadiobuttonGroupSelecRol.add(jRbtnStudent);
+        jRbtnStudent.setText("Estudiante");
+        jRbtnStudent.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jRbtnStudentActionPerformed(evt);
+            }
+        });
+        jPSelecRol.add(jRbtnStudent, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 40, -1, -1));
+
+        RadiobuttonGroupSelecRol.add(jRbtnCompany);
+        jRbtnCompany.setText("Empresa");
+        jRbtnCompany.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jRbtnCompanyActionPerformed(evt);
+            }
+        });
+        jPSelecRol.add(jRbtnCompany, new org.netbeans.lib.awtextra.AbsoluteConstraints(120, 40, -1, -1));
+
+        RadiobuttonGroupSelecRol.add(jRbtnCoordi);
+        jRbtnCoordi.setText("Coordinador");
+        jRbtnCoordi.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jRbtnCoordiActionPerformed(evt);
+            }
+        });
+        jPSelecRol.add(jRbtnCoordi, new org.netbeans.lib.awtextra.AbsoluteConstraints(220, 40, -1, -1));
+
+        jPContent.add(jPSelecRol, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 10, 360, 110));
+
+        jPRolInformation.setBackground(new java.awt.Color(255, 255, 255));
+        jPRolInformation.setForeground(new java.awt.Color(255, 255, 255));
+        jPRolInformation.setName(""); // NOI18N
+        jPRolInformation.setLayout(new java.awt.CardLayout());
+
+        jPCompanyInformation.setBackground(new java.awt.Color(255, 255, 255));
+        jPCompanyInformation.setName(""); // NOI18N
+
+        jLTittleNewCompany.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
+        jLTittleNewCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLTittleNewCompany.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLTittleNewCompany.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/NewUser.png"))); // NOI18N
+        jLTittleNewCompany.setText("Nueva Empresa");
+        jLTittleNewCompany.setHorizontalTextPosition(javax.swing.SwingConstants.RIGHT);
+
+        jTextAInfoCompany.setEditable(false);
+        jTextAInfoCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jTextAInfoCompany.setColumns(20);
+        jTextAInfoCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jTextAInfoCompany.setRows(5);
+        jTextAInfoCompany.setText("bla bla bla bla bla bla bla bla");
+        jScrolInfoCompany.setViewportView(jTextAInfoCompany);
+
+        javax.swing.GroupLayout jPCompanyInformationLayout = new javax.swing.GroupLayout(jPCompanyInformation);
+        jPCompanyInformation.setLayout(jPCompanyInformationLayout);
+        jPCompanyInformationLayout.setHorizontalGroup(
+            jPCompanyInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPCompanyInformationLayout.createSequentialGroup()
+                .addGroup(jPCompanyInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPCompanyInformationLayout.createSequentialGroup()
+                        .addGap(66, 66, 66)
+                        .addComponent(jLTittleNewCompany))
+                    .addGroup(jPCompanyInformationLayout.createSequentialGroup()
+                        .addGap(19, 19, 19)
+                        .addComponent(jScrolInfoCompany, javax.swing.GroupLayout.PREFERRED_SIZE, 317, javax.swing.GroupLayout.PREFERRED_SIZE)))
+                .addContainerGap(24, Short.MAX_VALUE))
+        );
+        jPCompanyInformationLayout.setVerticalGroup(
+            jPCompanyInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPCompanyInformationLayout.createSequentialGroup()
+                .addGap(14, 14, 14)
+                .addComponent(jLTittleNewCompany)
+                .addGap(18, 18, 18)
+                .addComponent(jScrolInfoCompany, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(40, Short.MAX_VALUE))
+        );
+
+        jPRolInformation.add(jPCompanyInformation, "CompanyInfo");
+
+        jPCoordinatorInformation.setBackground(new java.awt.Color(255, 255, 255));
+        jPCoordinatorInformation.setForeground(new java.awt.Color(255, 255, 255));
+        jPCoordinatorInformation.setName("CoordiInformation"); // NOI18N
+
+        jLTittleNewCompany1.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
+        jLTittleNewCompany1.setForeground(new java.awt.Color(0, 0, 0));
+        jLTittleNewCompany1.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLTittleNewCompany1.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/NewUser.png"))); // NOI18N
+        jLTittleNewCompany1.setText("Nuevo coordinador");
+        jLTittleNewCompany1.setHorizontalTextPosition(javax.swing.SwingConstants.RIGHT);
+
+        jTextAInfoCoordi.setEditable(false);
+        jTextAInfoCoordi.setBackground(new java.awt.Color(255, 255, 255));
+        jTextAInfoCoordi.setColumns(20);
+        jTextAInfoCoordi.setForeground(new java.awt.Color(0, 0, 0));
+        jTextAInfoCoordi.setRows(5);
+        jTextAInfoCoordi.setText("bla bla bla bla bla bla bla bla");
+        jScrolInfoCoordi.setViewportView(jTextAInfoCoordi);
+
+        javax.swing.GroupLayout jPCoordinatorInformationLayout = new javax.swing.GroupLayout(jPCoordinatorInformation);
+        jPCoordinatorInformation.setLayout(jPCoordinatorInformationLayout);
+        jPCoordinatorInformationLayout.setHorizontalGroup(
+            jPCoordinatorInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPCoordinatorInformationLayout.createSequentialGroup()
+                .addGroup(jPCoordinatorInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPCoordinatorInformationLayout.createSequentialGroup()
+                        .addGap(19, 19, 19)
+                        .addComponent(jScrolInfoCoordi, javax.swing.GroupLayout.PREFERRED_SIZE, 317, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(jPCoordinatorInformationLayout.createSequentialGroup()
+                        .addGap(52, 52, 52)
+                        .addComponent(jLTittleNewCompany1)))
+                .addContainerGap(24, Short.MAX_VALUE))
+        );
+        jPCoordinatorInformationLayout.setVerticalGroup(
+            jPCoordinatorInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPCoordinatorInformationLayout.createSequentialGroup()
+                .addGap(14, 14, 14)
+                .addComponent(jLTittleNewCompany1)
+                .addGap(18, 18, 18)
+                .addComponent(jScrolInfoCoordi, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(40, Short.MAX_VALUE))
+        );
+
+        jPRolInformation.add(jPCoordinatorInformation, "CoordiInfo");
+
+        jPStudentInformation.setBackground(new java.awt.Color(255, 255, 255));
+        jPStudentInformation.setName(""); // NOI18N
+
+        jLTittleNewStudent.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
+        jLTittleNewStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLTittleNewStudent.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLTittleNewStudent.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/NewUser.png"))); // NOI18N
+        jLTittleNewStudent.setText("Nuevo estudiante");
+        jLTittleNewStudent.setHorizontalTextPosition(javax.swing.SwingConstants.RIGHT);
+
+        jTextAInfoStudent.setEditable(false);
+        jTextAInfoStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jTextAInfoStudent.setColumns(20);
+        jTextAInfoStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jTextAInfoStudent.setRows(5);
+        jTextAInfoStudent.setText("bla bla bla bla bla bla bla bla");
+        jScrolInfoStudent.setViewportView(jTextAInfoStudent);
+
+        javax.swing.GroupLayout jPStudentInformationLayout = new javax.swing.GroupLayout(jPStudentInformation);
+        jPStudentInformation.setLayout(jPStudentInformationLayout);
+        jPStudentInformationLayout.setHorizontalGroup(
+            jPStudentInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPStudentInformationLayout.createSequentialGroup()
+                .addGroup(jPStudentInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+                    .addGroup(jPStudentInformationLayout.createSequentialGroup()
+                        .addGap(19, 19, 19)
+                        .addComponent(jScrolInfoStudent, javax.swing.GroupLayout.PREFERRED_SIZE, 317, javax.swing.GroupLayout.PREFERRED_SIZE))
+                    .addGroup(jPStudentInformationLayout.createSequentialGroup()
+                        .addGap(52, 52, 52)
+                        .addComponent(jLTittleNewStudent)))
+                .addContainerGap(24, Short.MAX_VALUE))
+        );
+        jPStudentInformationLayout.setVerticalGroup(
+            jPStudentInformationLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(jPStudentInformationLayout.createSequentialGroup()
+                .addGap(14, 14, 14)
+                .addComponent(jLTittleNewStudent)
+                .addGap(18, 18, 18)
+                .addComponent(jScrolInfoStudent, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(40, Short.MAX_VALUE))
+        );
+
+        jPRolInformation.add(jPStudentInformation, "StudentInfo");
+
+        jPContent.add(jPRolInformation, new org.netbeans.lib.awtextra.AbsoluteConstraints(20, 140, 360, 190));
+
+        jPRegisterUser.setBackground(new java.awt.Color(255, 255, 255));
+        jPRegisterUser.setLayout(new java.awt.CardLayout());
+
+        jPCoordiRegister.setBackground(new java.awt.Color(255, 204, 153));
+        jPCoordiRegister.setName("CoordiRegister"); // NOI18N
+
+        javax.swing.GroupLayout jPCoordiRegisterLayout = new javax.swing.GroupLayout(jPCoordiRegister);
+        jPCoordiRegister.setLayout(jPCoordiRegisterLayout);
+        jPCoordiRegisterLayout.setHorizontalGroup(
+            jPCoordiRegisterLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 410, Short.MAX_VALUE)
+        );
+        jPCoordiRegisterLayout.setVerticalGroup(
+            jPCoordiRegisterLayout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGap(0, 380, Short.MAX_VALUE)
+        );
+
+        jPRegisterUser.add(jPCoordiRegister, "CoordiRegis");
+
+        jPStudentRegister.setBackground(new java.awt.Color(255, 255, 255));
+        jPStudentRegister.setName("StudentRegister"); // NOI18N
+        java.awt.GridBagLayout jPStudentRegisterLayout = new java.awt.GridBagLayout();
+        jPStudentRegisterLayout.columnWidths = new int[] {200};
+        jPStudentRegister.setLayout(jPStudentRegisterLayout);
+
+        jLNameStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLNameStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLNameStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLNameStudent.setText("*Nombre:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 0;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 64;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(53, 99, 0, 0);
+        jPStudentRegister.add(jLNameStudent, gridBagConstraints);
+
+        jFieldNameStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldNameStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldNameStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldNameStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldNameStudent.setText("Julian Lopez");
+        jFieldNameStudent.setMaximumSize(new java.awt.Dimension(500, 100));
+        jFieldNameStudent.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFieldNameStudentActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 1;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jFieldNameStudent, gridBagConstraints);
+
+        jLCodeStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLCodeStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLCodeStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLCodeStudent.setText("Codigo:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 2;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 74;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jLCodeStudent, gridBagConstraints);
+
+        jFieldCodeStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldCodeStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldCodeStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldCodeStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldCodeStudent.setText("1046123");
+        jFieldCodeStudent.setMaximumSize(new java.awt.Dimension(500, 100));
+        jFieldCodeStudent.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFieldCodeStudentActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 3;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jFieldCodeStudent, gridBagConstraints);
+
+        jLTelStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLTelStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLTelStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLTelStudent.setText("Telefono de contacto:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 4;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jLTelStudent, gridBagConstraints);
+
+        jFieldTelStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldTelStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldTelStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldTelStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldTelStudent.setText("317123");
+        jFieldTelStudent.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 5;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jFieldTelStudent, gridBagConstraints);
+
+        jLEmailStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLEmailStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLEmailStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLEmailStudent.setText("Email:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 6;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 82;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jLEmailStudent, gridBagConstraints);
+
+        jFieldEmailStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldEmailStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldEmailStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldEmailStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldEmailStudent.setText("lopez@gmail.com");
+        jFieldEmailStudent.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 7;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jFieldEmailStudent, gridBagConstraints);
+
+        jLPasswordStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLPasswordStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jLPasswordStudent.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLPasswordStudent.setText("Clave:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 8;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 83;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPStudentRegister.add(jLPasswordStudent, gridBagConstraints);
+
+        jCheckSeePassStudent.setText("Mostrar clave");
+        jCheckSeePassStudent.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jCheckSeePassStudentActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 2;
+        gridBagConstraints.gridy = 9;
+        gridBagConstraints.gridheight = 2;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(10, 0, 0, 100);
+        jPStudentRegister.add(jCheckSeePassStudent, gridBagConstraints);
+
+        jFieldPassWordStudent.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldPassWordStudent.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldPassWordStudent.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldPassWordStudent.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldPassWordStudent.setText("jPasswordField1");
+        jFieldPassWordStudent.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 9;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(6, 99, 0, 0);
+        jPStudentRegister.add(jFieldPassWordStudent, gridBagConstraints);
+
+        jBtnSaveStudent.setBackground(new java.awt.Color(0, 0, 0));
+        jBtnSaveStudent.setFont(new java.awt.Font("SansSerif", 0, 14)); // NOI18N
+        jBtnSaveStudent.setForeground(new java.awt.Color(255, 255, 255));
+        jBtnSaveStudent.setText("GUARDAR");
+        jBtnSaveStudent.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtnSaveStudentActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 11;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(10, 108, 53, 0);
+        jPStudentRegister.add(jBtnSaveStudent, gridBagConstraints);
+
+        jPRegisterUser.add(jPStudentRegister, "StudentRegis");
+
+        jPCompanyRegister.setBackground(new java.awt.Color(255, 255, 255));
+        jPCompanyRegister.setForeground(new java.awt.Color(255, 255, 255));
+        jPCompanyRegister.setName("CompanyRegister"); // NOI18N
+        java.awt.GridBagLayout jPCompanyRegisterLayout = new java.awt.GridBagLayout();
+        jPCompanyRegisterLayout.columnWidths = new int[] {200};
+        jPCompanyRegister.setLayout(jPCompanyRegisterLayout);
+
+        jLNitCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLNitCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLNitCompany.setText("*NIT:");
+        jLNitCompany.setHorizontalTextPosition(javax.swing.SwingConstants.LEFT);
+        jLNitCompany.setVerticalTextPosition(javax.swing.SwingConstants.TOP);
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 0;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 90;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLNitCompany, gridBagConstraints);
+
+        jFieldNitCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldNitCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldNitCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldNitCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldNitCompany.setText("123456789");
+        jFieldNitCompany.setToolTipText("");
+        jFieldNitCompany.setAutoscrolls(false);
+        jFieldNitCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 1;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jFieldNitCompany, gridBagConstraints);
+
+        jLNameCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLNameCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLNameCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLNameCompany.setText("*Nombre Empresa");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 2;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 13;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLNameCompany, gridBagConstraints);
+
+        jFieldNameCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldNameCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldNameCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldNameCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldNameCompany.setText("Empresas Lopez");
+        jFieldNameCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        jFieldNameCompany.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFieldNameCompanyActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 3;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jFieldNameCompany, gridBagConstraints);
+
+        jLTelCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLTelCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLTelCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLTelCompany.setText("Telefono de contacto:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 4;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLTelCompany, gridBagConstraints);
+
+        jFieldTelCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldTelCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldTelCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldTelCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldTelCompany.setText("31232323");
+        jFieldTelCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        jFieldTelCompany.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jFieldTelCompanyActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 5;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jFieldTelCompany, gridBagConstraints);
+
+        jLWebCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLWebCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLWebCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLWebCompany.setText("Pagina web:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 6;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 49;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLWebCompany, gridBagConstraints);
+
+        jFieldWebCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldWebCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldWebCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldWebCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldWebCompany.setText("www.lopez.com");
+        jFieldWebCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 7;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jFieldWebCompany, gridBagConstraints);
+
+        jLSectorCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLSectorCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLSectorCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLSectorCompany.setText("Sector industrial:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 8;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 25;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLSectorCompany, gridBagConstraints);
+
+        jCBSectorCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jCBSectorCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jCBSectorCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jCBSectorCompany.setModel(new javax.swing.DefaultComboBoxModel<>(new String[] { "Item 1", "Item 2", "Item 3", "Item 4" }));
+        jCBSectorCompany.setMaximumSize(new java.awt.Dimension(300, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 9;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 45;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jCBSectorCompany, gridBagConstraints);
+
+        jLEmailCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLEmailCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLEmailCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLEmailCompany.setText("Email:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 10;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 82;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLEmailCompany, gridBagConstraints);
+
+        jFieldEmailCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldEmailCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldEmailCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldEmailCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldEmailCompany.setText("lopez@gmail.com");
+        jFieldEmailCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 11;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jFieldEmailCompany, gridBagConstraints);
+
+        jLPasswordCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jLPasswordCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jLPasswordCompany.setHorizontalAlignment(javax.swing.SwingConstants.LEFT);
+        jLPasswordCompany.setText("Clave:");
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 12;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 83;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(5, 99, 0, 0);
+        jPCompanyRegister.add(jLPasswordCompany, gridBagConstraints);
+
+        jFieldPassWordCompany.setBackground(new java.awt.Color(255, 255, 255));
+        jFieldPassWordCompany.setFont(new java.awt.Font("SansSerif", 0, 12)); // NOI18N
+        jFieldPassWordCompany.setForeground(new java.awt.Color(0, 0, 0));
+        jFieldPassWordCompany.setHorizontalAlignment(javax.swing.JTextField.LEFT);
+        jFieldPassWordCompany.setText("Clave123.");
+        jFieldPassWordCompany.setMaximumSize(new java.awt.Dimension(500, 100));
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 13;
+        gridBagConstraints.gridwidth = 2;
+        gridBagConstraints.fill = java.awt.GridBagConstraints.HORIZONTAL;
+        gridBagConstraints.ipadx = 53;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(6, 99, 0, 0);
+        jPCompanyRegister.add(jFieldPassWordCompany, gridBagConstraints);
+
+        jCheckSeePassCompany.setText("Mostrar clave");
+        jCheckSeePassCompany.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jCheckSeePassCompanyActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 2;
+        gridBagConstraints.gridy = 13;
+        gridBagConstraints.gridheight = 2;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(10, 0, 0, 100);
+        jPCompanyRegister.add(jCheckSeePassCompany, gridBagConstraints);
+
+        jBtnSaveCompany.setBackground(new java.awt.Color(0, 0, 0));
+        jBtnSaveCompany.setFont(new java.awt.Font("SansSerif", 0, 14)); // NOI18N
+        jBtnSaveCompany.setForeground(new java.awt.Color(255, 255, 255));
+        jBtnSaveCompany.setText("GUARDAR");
+        jBtnSaveCompany.addActionListener(new java.awt.event.ActionListener() {
+            public void actionPerformed(java.awt.event.ActionEvent evt) {
+                jBtnSaveCompanyActionPerformed(evt);
+            }
+        });
+        gridBagConstraints = new java.awt.GridBagConstraints();
+        gridBagConstraints.gridx = 0;
+        gridBagConstraints.gridy = 15;
+        gridBagConstraints.anchor = java.awt.GridBagConstraints.NORTHWEST;
+        gridBagConstraints.insets = new java.awt.Insets(10, 108, 5, 0);
+        jPCompanyRegister.add(jBtnSaveCompany, gridBagConstraints);
+
+        jPRegisterUser.add(jPCompanyRegister, "CompanyRegis");
+
+        jPContent.add(jPRegisterUser, new org.netbeans.lib.awtextra.AbsoluteConstraints(390, 10, 410, 380));
+
+        jPTittleNewUser.setBackground(new java.awt.Color(255, 255, 255));
+
+        jLTittleNewUser.setFont(new java.awt.Font("SansSerif", 1, 24)); // NOI18N
+        jLTittleNewUser.setForeground(new java.awt.Color(0, 0, 0));
+        jLTittleNewUser.setHorizontalAlignment(javax.swing.SwingConstants.CENTER);
+        jLTittleNewUser.setIcon(new javax.swing.ImageIcon(getClass().getResource("/imagenes/NewUser.png"))); // NOI18N
+        jLTittleNewUser.setText("Nuevo Registro ");
+        jLTittleNewUser.setHorizontalTextPosition(javax.swing.SwingConstants.RIGHT);
+        jPTittleNewUser.add(jLTittleNewUser);
+
+        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
+        getContentPane().setLayout(layout);
+        layout.setHorizontalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addComponent(jPTittleNewUser, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(jPHead, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, 808, Short.MAX_VALUE)
+            .addComponent(jPButtom, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(jPContent, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+        );
+        layout.setVerticalGroup(
+            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
+            .addGroup(layout.createSequentialGroup()
+                .addComponent(jPHead, javax.swing.GroupLayout.PREFERRED_SIZE, 101, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPTittleNewUser, javax.swing.GroupLayout.PREFERRED_SIZE, 63, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPContent, javax.swing.GroupLayout.PREFERRED_SIZE, 397, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED)
+                .addComponent(jPButtom, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
+        );
+
+        pack();
+    }// </editor-fold>//GEN-END:initComponents
+
+    private void jFieldNameCompanyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldNameCompanyActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jFieldNameCompanyActionPerformed
+
+    private void jFieldTelCompanyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldTelCompanyActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jFieldTelCompanyActionPerformed
+
+    private void jCheckSeePassCompanyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCheckSeePassCompanyActionPerformed
+        if (jCheckSeePassCompany.isSelected()) {
+            jFieldPassWordCompany.setEchoChar((char) 0); // Mostrar la contraseña
+        } else {
+            jFieldPassWordCompany.setEchoChar('●'); // Ocultar la contraseña con puntos
+        }
+    }//GEN-LAST:event_jCheckSeePassCompanyActionPerformed
+
+    private void jBtnSaveCompanyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtnSaveCompanyActionPerformed
+        // TODO add your handling code here:
+        validateCompany validar = new validateCompany();
+
+        String nit = jFieldNitCompany.getText();
+        if (nit.equals("")) {
+            Messages.showMessageDialog("Debe agregar el Nit", "Atención");
+            jFieldNitCompany.requestFocus();
+            return;
+        }
+
+        String nombre = jFieldNameCompany.getText();
+        if (nombre.equals("")) {
+            Messages.showMessageDialog("Debe agregar el Nombre", "Atención");
+            jFieldNameCompany.requestFocus();
+            return;
+        }
+
+        String telefono = jFieldTelCompany.getText();
+        String pagina = jFieldWebCompany.getText();
+        Sector sector = Sector.values()[jCBSectorCompany.getSelectedIndex()];
+
+        String email = jFieldEmailCompany.getText().trim();
+        if (email.equals("")) {
+            Messages.showMessageDialog("Debe agregar el Email", "Atención");
+            jFieldEmailCompany.requestFocus();
+            return;
+        }
+
+        String pass = jFieldPassWordCompany.getText().trim();
+        if (pass.equals("")) {
+            Messages.showMessageDialog("Debe agregar la Contraseña", "Atención");
+            jFieldPassWordCompany.requestFocus();
+            return;
+        }
+        if (!validar.validarClave(pass)) {
+            validar.mostrarEjemploClave();
+        }
+        if (validar.validarRegistroEmpresa(nit, email, telefono, pass)) {
+
+            if (companyService.saveCompany(new Company(nit, nombre, telefono, pagina, sector, email, pass))) {
+                GUIHomeCompany hCompany = new GUIHomeCompany();
+                hCompany.setVisible(true);
+                this.dispose();
+            }
+        }
+    }//GEN-LAST:event_jBtnSaveCompanyActionPerformed
+
+    private void jRbtnStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jRbtnStudentActionPerformed
+        SeleccionarRol();
+    }//GEN-LAST:event_jRbtnStudentActionPerformed
+
+    private void jRbtnCompanyActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jRbtnCompanyActionPerformed
+        SeleccionarRol();
+    }//GEN-LAST:event_jRbtnCompanyActionPerformed
+
+    private void jRbtnCoordiActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jRbtnCoordiActionPerformed
+        SeleccionarRol();
+    }//GEN-LAST:event_jRbtnCoordiActionPerformed
+
+    private void jFieldNameStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldNameStudentActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jFieldNameStudentActionPerformed
+
+    private void jFieldCodeStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jFieldCodeStudentActionPerformed
+        // TODO add your handling code here:
+    }//GEN-LAST:event_jFieldCodeStudentActionPerformed
+
+    private void jCheckSeePassStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jCheckSeePassStudentActionPerformed
+        // TODO add your handling code here:
+        if (jCheckSeePassStudent.isSelected()) {
+            jFieldPassWordStudent.setEchoChar((char) 0); // Mostrar la contraseña
+        } else {
+            jFieldPassWordStudent.setEchoChar('*'); // Ocultar la contraseña con puntos
+        }
+    }//GEN-LAST:event_jCheckSeePassStudentActionPerformed
+
+    private void jBtnSaveStudentActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jBtnSaveStudentActionPerformed
+        // TODO add your handling code here:
+        validateStudent validar = new validateStudent();
+
+        String name = jFieldNameStudent.getText();
+        if (name.equals("")) {
+            Messages.showMessageDialog("Debe agregar el nombre", "Atención");
+            jFieldNameStudent.requestFocus();
+            return;
+        }
+
+        String code = jFieldCodeStudent.getText();
+        if (code.equals("")) {
+            Messages.showMessageDialog("Debe agregar el codigo", "Atención");
+            jFieldCodeStudent.requestFocus();
+            return;
+        }
+
+        String telefono = jFieldTelStudent.getText();
+
+        String email = jFieldEmailStudent.getText().trim();
+        if (email.equals("")) {
+            Messages.showMessageDialog("Debe agregar el Email", "Atención");
+            jFieldEmailStudent.requestFocus();
+            return;
+        }
+
+        String pass = jFieldPassWordStudent.getText().trim();
+        if (pass.equals("")) {
+            Messages.showMessageDialog("Debe agregar la Contraseña", "Atención");
+            jFieldPassWordStudent.requestFocus();
+            return;
+        }
+        if (!validar.validarClave(pass)) {
+            validar.mostrarEjemploClave();
+        }
+        if (validar.validarRegistroEstudiante(email, telefono, pass)) {
+
+            if (studentService.save(new Student(code, name, telefono, email, pass))) {
+                GUIHomeStudent hStudent = new GUIHomeStudent();
+                hStudent.setVisible(true);
+                this.dispose();
+            }
+
+        }
+    }//GEN-LAST:event_jBtnSaveStudentActionPerformed
+
+    private void SeleccionarRol() {
+        if (jRbtnCompany.isSelected()) {
+            cardLayout = (CardLayout) jPRolInformation.getLayout();
+            cardLayout.show(jPRolInformation, "CompanyInfo");
+            cardLayout = (CardLayout) jPRegisterUser.getLayout();
+            cardLayout.show(jPRegisterUser, "CompanyRegis");
+        }
+        if (jRbtnStudent.isSelected()) {
+            cardLayout = (CardLayout) jPRolInformation.getLayout();
+            cardLayout.show(jPRolInformation, "StudentInfo");
+            cardLayout = (CardLayout) jPRegisterUser.getLayout();
+            cardLayout.show(jPRegisterUser, "StudentRegis");
+        }
+        if (jRbtnCoordi.isSelected()) {
+            cardLayout = (CardLayout) jPRolInformation.getLayout();
+            cardLayout.show(jPRolInformation, "CoordinatorInfo");
+            cardLayout = (CardLayout) jPRegisterUser.getLayout();
+            cardLayout.show(jPRegisterUser, "CoordiRegis");
+        }
+    }
+
+
+    // Variables declaration - do not modify//GEN-BEGIN:variables
+    private javax.swing.ButtonGroup RadiobuttonGroupSelecRol;
+    private javax.swing.JButton jBtnBackSelectUser;
+    private javax.swing.JButton jBtnSaveCompany;
+    private javax.swing.JButton jBtnSaveStudent;
+    private javax.swing.JComboBox<String> jCBSectorCompany;
+    private javax.swing.JCheckBox jCheckSeePassCompany;
+    private javax.swing.JCheckBox jCheckSeePassStudent;
+    private javax.swing.JTextField jFieldCodeStudent;
+    private javax.swing.JTextField jFieldEmailCompany;
+    private javax.swing.JTextField jFieldEmailStudent;
+    private javax.swing.JTextField jFieldNameCompany;
+    private javax.swing.JTextField jFieldNameStudent;
+    private javax.swing.JTextField jFieldNitCompany;
+    private javax.swing.JPasswordField jFieldPassWordCompany;
+    private javax.swing.JPasswordField jFieldPassWordStudent;
+    private javax.swing.JTextField jFieldTelCompany;
+    private javax.swing.JTextField jFieldTelStudent;
+    private javax.swing.JTextField jFieldWebCompany;
+    private javax.swing.JLabel jLCodeStudent;
+    private javax.swing.JLabel jLEmailCompany;
+    private javax.swing.JLabel jLEmailStudent;
+    private javax.swing.JLabel jLNameCompany;
+    private javax.swing.JLabel jLNameStudent;
+    private javax.swing.JLabel jLNitCompany;
+    private javax.swing.JLabel jLPasswordCompany;
+    private javax.swing.JLabel jLPasswordStudent;
+    private javax.swing.JLabel jLSectorCompany;
+    private javax.swing.JLabel jLSelectRol;
+    private javax.swing.JLabel jLTelCompany;
+    private javax.swing.JLabel jLTelStudent;
+    private javax.swing.JLabel jLTittleNewCompany;
+    private javax.swing.JLabel jLTittleNewCompany1;
+    private javax.swing.JLabel jLTittleNewStudent;
+    private javax.swing.JLabel jLTittleNewUser;
+    private javax.swing.JLabel jLWebCompany;
+    private javax.swing.JPanel jPButtom;
+    private javax.swing.JPanel jPCompanyInformation;
+    private javax.swing.JPanel jPCompanyRegister;
+    private javax.swing.JPanel jPContent;
+    private javax.swing.JPanel jPCoordiRegister;
+    private javax.swing.JPanel jPCoordinatorInformation;
+    private javax.swing.JPanel jPHead;
+    private javax.swing.JPanel jPRegisterUser;
+    private javax.swing.JPanel jPRolInformation;
+    private javax.swing.JPanel jPSelecRol;
+    private javax.swing.JPanel jPStudentInformation;
+    private javax.swing.JPanel jPStudentRegister;
+    private javax.swing.JPanel jPTittleNewUser;
+    private javax.swing.JRadioButton jRbtnCompany;
+    private javax.swing.JRadioButton jRbtnCoordi;
+    private javax.swing.JRadioButton jRbtnStudent;
+    private javax.swing.JScrollPane jScrolInfoCompany;
+    private javax.swing.JScrollPane jScrolInfoCoordi;
+    private javax.swing.JScrollPane jScrolInfoStudent;
+    private javax.swing.JTextArea jTextAInfoCompany;
+    private javax.swing.JTextArea jTextAInfoCoordi;
+    private javax.swing.JTextArea jTextAInfoStudent;
+    private java.awt.Label lbTitleProyect;
+    // End of variables declaration//GEN-END:variables
+}
